1920582: Added additional validation for cloud registration data
- Added an additional check for data in the /cloud/authorize endpoint
  to trigger a 400 rather than a 500 when no cloud registration
  details are provided
- Fixed a bug with the serialization of exception output when writing
  out a CandlepinException that does not have an exception message
- Candlepin will now list "cloud_registration" as a capability when
  the cloud registration feature is enabled
- Added support for cloud registration to the spec tests
- Added additional testing surrounding the cloud registration feature
- Other minor improvements and optimizations to the spec tests
